XXX-XXX: Update regex condition for Status in EsbService
Enhanced the conditional statement in the Common.Services namespace of EsbService.cs by adding a new regular expression to check for various formats of the Status element in the RE.RESPOSTA string. This change improves the flexibility of the condition while maintaining the overall structure of the whereClause for SQL queries.